Added Speeds

+++ include/TimeUtil.h
@@ -3,7 +3,7 @@
 class TimeUtil
 {
 	public:
-		static float TvDiffMs(const struct timeval &tv1, const struct timeval &tv2);
+		static double TvDiffMs(const struct timeval &tv1, const struct timeval &tv2);
 };
